Minor fixes

+++ src/TestScript.java
@@ -37,7 +37,7 @@ public class TestScript {
         System.out.println(">> TEST Parsing Simple Script");
         var bos = new ByteArrayOutputStream();
         //bos.write(3);
-        bos.write(CryptoKit.encodeVarint((long)3));
+        bos.write(CryptoKit.encodeVarint(3));
         bos.write(0x54);
         bos.write(0);
         bos.write(79);
